added testing on ccl

+++ t/binding.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
+run_test (){
+    if which $1
+    then
+        $@ binding.lisp &
+        pid=$!
+        sleep 15
+        echo "killing $pid with -9"
+        kill -9 $pid
+    else
+        echo "$1 not installed"
+    fi
+}
 
-sbcl --load binding.lisp &
-pid=$!
-sleep 30
-echo "killing $pid with -9"
-kill -9 $pid
-
+run_test sbcl --quit --load
+run_test ccl --batch --load
